third_party/blink/web_tests/wpt_internal/compression-dictionary/fetch-from-link-element.https.html

<!DOCTYPE html>
<html>
<head>
  <title>Load compression dictionary from link element</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/common/get-host-info.sub.js"></script>
  <script src="/common/utils.js"></script>
  <script src="resources/dict_fetch_test.js"></script>
  <script src="resources/compression-dictionary-origin-trial.js"></script>
</head>
<body>
  <script>
    enableCompressionDictionaryOriginTrial()

    function setup_dict_link(token, remote_origin) {
      const dictionary_link = document.createElement('link');
      dictionary_link.rel = 'dictionary';
      const same_origin_dict =
        '/wpt_internal/compression-dictionary/resources/dict.py?token=' + token;
      if (remote_origin) {
        dictionary_link.href = remote_origin + same_origin_dict;
        dictionary_link.crossOrigin = '';
      } else {
        dictionary_link.href = same_origin_dict
      }
      document.body.appendChild(dictionary_link);
    }

    async function fetch_previous_dictionary_load_header(token) {
      await new Promise(resolve => window.requestIdleCallback(resolve));
      const response = await fetch('resources/dict.py?token=' + token + '&previous_header=1');
      return response.json();
    }

    promise_test(async test => {
      const dict_token = token();
      setup_dict_link(dict_token, null);
      const dict_fetch_response = await fetch_previous_dictionary_load_header(dict_token);
      checkSameOriginDictionaryFetchHeaders(dict_fetch_response);
    }, 'Dictionary correctly fetched');

    promise_test(async test => {
      const dict_token = token();
      setup_dict_link(dict_token, get_host_info().HTTPS_NOTSAMESITE_ORIGIN);
      const dict_fetch_response = await fetch_previous_dictionary_load_header(dict_token);
      checkNotSameSiteDictionaryFetchHeaders(dict_fetch_response);
    }, 'Not-same-site origin dictionary correctly fetched');

    promise_test(async test => {
      const dict_token = token();
      setup_dict_link(dict_token, get_host_info().HTTPS_REMOTE_ORIGIN);
      const dict_fetch_response = await fetch_previous_dictionary_load_header(dict_token);
      checkRemoteOriginDictionaryFetchHeaders(dict_fetch_response);
    }, 'Remote origin dictionary correctly fetched');

  </script>
</body>
</html>
